+++ src/features/reservation/reservation-list/ui/my-reservation-list/ReservationListView.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react';
 import {
   MY_RESERVATIONS_PAGE_SIZE,
   useMyReservations,
-} from '@/features/reservation/reservation-list/queries/useMyReservations';
+} from '@/features/reservation/reservation-list/queries/useInfiniteMyReservations';
 import ReservationListSection from '@/features/reservation/reservation-list/ui/my-reservation-list/ReservationListSection';
 import StatusFilter from '@/features/reservation/reservation-list/ui/status-filter/StatusFilter';
 import type { MyReservationStatus } from '@/features/reservation/types';
